fix(plugin): Optional cleanup

+++ src/io/stenic/jpipe/plugin/DockerPlugin.groovy
@@ -13,6 +13,7 @@ class DockerPlugin extends Plugin {
     private String testScript;
     private List extraTargets;
     private Boolean useCache;
+    private Boolean doCleanup;
 
     DockerPlugin(Map opts = [:]) {
         this.repository = opts.get('repository', '');
@@ -24,6 +25,7 @@ class DockerPlugin extends Plugin {
         this.extraTargets = opts.get('extraTargets', []);
         this.testScript = opts.get('testScript', '');
         this.useCache = opts.get('useCache', true);
+        this.doCleanup = opts.get('doCleanup', false);
     }
 
     public Map getSubscribedEvents() {
@@ -90,12 +92,19 @@ class DockerPlugin extends Plugin {
     }
 
     public void doDockerCleanup(Event event) {
+        if (!this.doCleanup) {
+            return
+        }
         try {
             event.script.sh "docker rmi ${this.repository}:${event.version}"
-            event.script.sh "docker rmi ${this.repository}:cache"
             this.extraTargets.each { target ->
                 event.script.sh "docker rmi ${this.repository}:${target}"
-                event.script.sh "docker rmi ${this.repository}:cache-${target}"
+            }
+            if (this.useCache) {
+                event.script.sh "docker rmi ${this.repository}:cache"
+                this.extraTargets.each { target ->
+                    event.script.sh "docker rmi ${this.repository}:cache-${target}"
+                }
             }
             event.script.sh 'docker rmi -f $(docker images -f "dangling=true" -q)'
         } catch(Exception e) {}
